Fixed bug with disabling GL_TEXTURE_2D in FTTextureFont.
FTTextureFont::RenderI was setting two state tracking variables
disableTexture2D and disableBlend. These are also set in
FTTextureFont::PreRender(). When we have to call RenderI multiple
times (such as a multi line string), we only call
PreRender/PostRender twice, but since Render was resetting these
two state tokens, we'd forget what the real state was, so PostRender
was not restoring the correct state.