Added Set/GetLineHeight(float) to FTSimpleFontLayout.
SetLineHeight(float) allows you to set a layouts line height in GL units,
as opposed to SetLineSpacing which was only a percentage of the fonts
line height.

With SetLineHeight you can align your text onto a baseline grid,
instead of relying on whatever the font was using for a line
height and trying to match some division of that to your baseline.